Modify sample for framework
