+++ YggdrAshill.Ragnarok.Samples/YggdrAshill.Ragnarok.Samples.csproj
@@ -3,12 +3,9 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <AssemblyName>YggdrAshill.Ragnarok.Samples</AssemblyName>
-    <RootNamespace>YggdrAshill.Ragnarok.Samples</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Progression\Progression.csproj" />
     <ProjectReference Include="..\YggdrAshill.Ragnarok\YggdrAshill.Ragnarok.csproj" />
   </ItemGroup>
