+++ YggdrAshill.Ragnarok.sln
@@ -9,12 +9,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YggdrAshill.Ragnarok.Specif
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Progression", "Progression\Progression.csproj", "{3B10218B-3A4B-418E-B68A-E678187E50ED}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Samples", "Samples\Samples.csproj", "{07228844-E269-4EB9-B0DA-C7353EABB9BF}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9E682B21-AA54-4CB7-AB23-5EB93BAFC78C}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YggdrAshill.Ragnarok.Samples", "YggdrAshill.Ragnarok.Samples\YggdrAshill.Ragnarok.Samples.csproj", "{BF4F0620-69F6-417D-9A20-7D1E164890AB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,10 +34,10 @@ Global
 		{3B10218B-3A4B-418E-B68A-E678187E50ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B10218B-3A4B-418E-B68A-E678187E50ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3B10218B-3A4B-418E-B68A-E678187E50ED}.Release|Any CPU.Build.0 = Release|Any CPU
-		{07228844-E269-4EB9-B0DA-C7353EABB9BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{07228844-E269-4EB9-B0DA-C7353EABB9BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{07228844-E269-4EB9-B0DA-C7353EABB9BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{07228844-E269-4EB9-B0DA-C7353EABB9BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF4F0620-69F6-417D-9A20-7D1E164890AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF4F0620-69F6-417D-9A20-7D1E164890AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF4F0620-69F6-417D-9A20-7D1E164890AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF4F0620-69F6-417D-9A20-7D1E164890AB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
